Account creation and sign in should all be good
set up server and made the first frame for creating an account or signing into an account. Deleted the need for the users username because we never used it and it only makes the account creation look more sloppy. If we kept it, I would probably have to make a new frame for creating an account which would be annoying.