Half Functional Add to Cart
In the case where you add the same item to your cart multiple times, it doesn't increment on the existing cart item, but just appends new cart item with new quantity, in addition to the already existing cart item. This has quite a few implications that break the program outright, but for now it kind of works.